Install: set innodb_file_per_table=1 to mysql conf
In MySQL 5.5 (current version in Debian Jessie used by jeedom), innodb_file_per_table is not enabled by default.
It allows to have a .ibd file per table, allowing easier restoration in case of corruption. Since jeedom is heavily used on device using sdcards, plaggued by corruption problems, it should be a good idea to set it.

This commit is to be reverted as soon as jeedom recommends using a version of debian shipping with MySQL 5.7 (or, better, MariaDB?).