throw exception instead of returning obscure error codes
When using these functions, for example in the expression evaluator, we get an exception if the command is not found (which is great), but -1 or -2 if the command is not historized or its history is empty. I think we should throw different exceptions for all these cases, unless the return value is actually used.