Unify apt operations during installation
This commit aims at reducing the number of apt invokations to save time.
On my machine, the optimization yields:
  step_2: 6'30 before, 5' after (75%).
  step_4: 3' before, 2' after (66%).

This commit also stops trying installing libav-tools and mbrola, as apt
cannot install them. It also stops trying removing brltty and net-tools,
as they are not installed. This was inferred from the following logs:

---
Package libav-tools is not available, but is referred to by another package.
This may mean that the package is missing, has been obsoleted, or
is only available from another source
However the following packages replace it:
  ffmpeg

Package mbrola is not available, but is referred to by another package.
This may mean that the package is missing, has been obsoleted, or
is only available from another source

E: Package 'libav-tools' has no installation candidate
E: Package 'mbrola' has no installation candidate
Reading package lists...
Building dependency tree...
Reading state information...
Package 'brltty' is not installed, so not removed
Package 'net-tools' is not installed, so not removed
---