Add unset($component['type']);
Add unset($component['type']); before a2o
the 'type' is not used in a2o, but get trouble if a function is called "settype" in a plugin. (get 'true' with method_exists gettype)
For example, the jMQTT plugin uses the setType() setter, when a user change a view (with the pencil), it changes the type by eqLogic in the plugin.

This modification is a prevention in case another plugin uses the same settler.